refactor(api): move copy-organization target profile into bounded context
