+++ api/src/prescription/target-profile/domain/usecases/attach-organizations-from-existing-target-profile.js
@@ -1,4 +1,4 @@
-import { OrganizationsToAttachToTargetProfile } from '../models/index.js';
+import { OrganizationsToAttachToTargetProfile } from '../../../../../lib/domain/models/index.js';
 
 const attachOrganizationsFromExistingTargetProfile = async function ({
   targetProfileId,
