+++ api/tests/integration/domain/usecases/attach-organizations-to-target-profile_test.js
@@ -1,5 +1,5 @@
 import { attachOrganizationsToTargetProfile } from '../../../../lib/domain/usecases/attach-organizations-to-target-profile.js';
-import * as organizationsToAttachToTargetProfileRepository from '../../../../lib/infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
+import * as organizationsToAttachToTargetProfileRepository from '../../../../src/prescription/target-profile/infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
 import { skillDatasource } from '../../../../src/shared/infrastructure/datasources/learning-content/skill-datasource.js';
 import { databaseBuilder, expect, knex, sinon } from '../../../test-helper.js';
