+++ api/tests/prescription/target-profile/integration/domain/attach-organizations-from-existing-target-profile_test.js
@@ -1,10 +1,10 @@
-import { NoOrganizationToAttach } from '../../../../lib/domain/errors.js';
-import { attachOrganizationsFromExistingTargetProfile } from '../../../../lib/domain/usecases/attach-organizations-from-existing-target-profile.js';
-import * as organizationsToAttachToTargetProfileRepository from '../../../../lib/infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
-import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { skillDatasource } from '../../../../src/shared/infrastructure/datasources/learning-content/skill-datasource.js';
-import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../test-helper.js';
+import { NoOrganizationToAttach } from '../../../../../lib/domain/errors.js';
+import * as targetProfileRepository from '../../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import { attachOrganizationsFromExistingTargetProfile } from '../../../../../src/prescription/target-profile/domain/usecases/attach-organizations-from-existing-target-profile.js';
+import * as organizationsToAttachToTargetProfileRepository from '../../../../../src/prescription/target-profile/infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
+import { skillDatasource } from '../../../../../src/shared/infrastructure/datasources/learning-content/skill-datasource.js';
+import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
 
 describe('Integration | UseCase | attach-organizations-from-existing-target-profile', function () {
   beforeEach(function () {
