+++ api/tests/prescription/target-profile/unit/application/admin-target-profiles-route_test.js
@@ -11,6 +11,96 @@ describe('Unit | Application | Target Profiles | Routes', function () {
     sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf');
   });
 
+  describe('POST /api/admin/target-profiles/{id}/copy-organizations', function () {
+    const method = 'POST';
+    const url = '/api/admin/target-profiles/3/copy-organizations';
+    const payload = { 'target-profile-id': 1 };
+
+    context('when user has role "SUPER_ADMIN", "SUPPORT" or "METIER"', function () {
+      it('should return a response with an HTTP status code 204', async function () {
+        // given
+        securityPreHandlers.hasAtLeastOneAccessOf
+          .withArgs([
+            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+            securityPreHandlers.checkAdminMemberHasRoleSupport,
+            securityPreHandlers.checkAdminMemberHasRoleMetier,
+          ])
+          .callsFake(() => (request, h) => h.response(true));
+        sinon
+          .stub(targetProfileController, 'attachOrganizationsFromExistingTargetProfile')
+          .callsFake((request, h) => h.response('ok').code(204));
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const { statusCode } = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(statusCode).to.equal(204);
+        sinon.assert.calledOnce(targetProfileController.attachOrganizationsFromExistingTargetProfile);
+      });
+
+      context('when id is a string', function () {
+        it('should reject request with HTTP code 400', async function () {
+          // given
+          const httpTestServer = new HttpTestServer();
+          await httpTestServer.register(moduleUnderTest);
+
+          // when
+          const { statusCode } = await httpTestServer.request(
+            method,
+            '/api/admin/target-profiles/azerty/copy-organizations',
+            payload,
+          );
+
+          // then
+          expect(statusCode).to.equal(400);
+        });
+      });
+
+      context('when target profile id is a string', function () {
+        it('should reject request with HTTP code 400', async function () {
+          // given
+          const httpTestServer = new HttpTestServer();
+          await httpTestServer.register(moduleUnderTest);
+
+          // when
+          const { statusCode } = await httpTestServer.request(method, url, { 'target-profile-id': 'azerty' });
+
+          // then
+          expect(statusCode).to.equal(400);
+        });
+      });
+    });
+
+    context('when user has role "CERTIF"', function () {
+      it('should return a response with an HTTP status code 403', async function () {
+        // given
+        securityPreHandlers.hasAtLeastOneAccessOf
+          .withArgs([
+            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+            securityPreHandlers.checkAdminMemberHasRoleSupport,
+            securityPreHandlers.checkAdminMemberHasRoleMetier,
+          ])
+          .callsFake(
+            () => (request, h) =>
+              h
+                .response({ errors: new Error('forbidden') })
+                .code(403)
+                .takeover(),
+          );
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const { statusCode } = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(statusCode).to.equal(403);
+      });
+    });
+  });
+
   describe('GET /api/admin/target-profiles/{id}/learning-content-pdf?language={language}', function () {
     let method, httpTestServer;
