+++ api/tests/prescription/target-profile/unit/infrastructure/serializers/jsonapi/target-profile-attach-organization-serializer_test.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/target-profile-attach-organization-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../../lib/infrastructure/serializers/jsonapi/target-profile-attach-organization-serializer.js';
+import { expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | target-profile-attach-organization-serializer', function () {
   describe('#serialize', function () {
